+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 * bumped dirs to 3.0
 * only allow 'c' and 'r' bindings to do things if app is not paused and is currently on a task
 * hanging back on rodio 0.11 until [this](https://github.com/RustAudio/rodio/issues/290) is resolved
+* add 'h' to toggle a help menu which describes keypresses
+* ROUNDED CORNERS!!!